+++ sql-plugin/src/main/spark341db/scala/com/nvidia/spark/rapids/shims/Spark341PlusDBShims.scala
@@ -28,40 +28,21 @@ import org.apache.spark.sql.catalyst.plans.physical.SinglePartition
 import org.apache.spark.sql.execution.{CollectLimitExec, GlobalLimitExec, SparkPlan, TakeOrderedAndProjectExec}
 import org.apache.spark.sql.execution.exchange.ENSURE_REQUIREMENTS
 import org.apache.spark.sql.rapids.GpuV1WriteUtils.GpuEmpty2Null
-import org.apache.spark.sql.rapids.execution.python.GpuPythonUDAF
 
 trait Spark341PlusDBShims extends Spark332PlusDBShims {
 
   override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = {
     val shimExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = Seq(
-      GpuOverrides.expr[PythonUDAF](
-        "UDF run in an external python process. Does not actually run on the GPU, but " +
-          "the transfer of data to/from it can be accelerated",
-        ExprChecks.fullAggAndProject(
-          // Different types of Pandas UDF support different sets of output type. Please refer to
-          //   https://github.com/apache/spark/blob/master/python/pyspark/sql/udf.py#L98
-          // for more details.
-          // It is impossible to specify the exact type signature for each Pandas UDF type in a
-          // single expression 'PythonUDF'.
-          // So use the 'unionOfPandasUdfOut' to cover all types for Spark. The type signature of
-          // plugin is also an union of all the types of Pandas UDF.
-          (TypeSig.commonCudfTypes + TypeSig.ARRAY).nested() + TypeSig.STRUCT,
-          TypeSig.unionOfPandasUdfOut,
-          repeatingParamCheck = Some(RepeatingParamCheck(
-            "param",
-            (TypeSig.commonCudfTypes + TypeSig.ARRAY + TypeSig.STRUCT).nested(),
-            TypeSig.all))),
-        (a, conf, p, r) => new ExprMeta[PythonUDAF](a, conf, p, r) {
-          override def replaceMessage: String = "not block GPU acceleration"
-
-          override def noReplacementPossibleMessage(reasons: String): String =
-            s"blocks running on GPU because $reasons"
-
-          override def convertToGpu(): GpuExpression =
-            GpuPythonUDAF(a.name, a.func, a.dataType,
-              childExprs.map(_.convertToGpu()),
-              a.evalType, a.udfDeterministic, a.resultId)
-        })
+      // Empty2Null is pulled out of FileFormatWriter by default since Spark 3.4.0,
+      // so it is visible in the overriding stage.
+      GpuOverrides.expr[Empty2Null](
+        "Converts the empty string to null for writing data",
+        ExprChecks.unaryProjectInputMatchesOutput(
+          TypeSig.STRING, TypeSig.STRING),
+        (a, conf, p, r) => new UnaryExprMeta[Empty2Null](a, conf, p, r) {
+          override def convertToGpu(child: Expression): GpuExpression = GpuEmpty2Null(child)
+        }
+      )
     ).map(r => (r.getClassFor.asSubclass(classOf[Expression]), r)).toMap
     super.getExprs ++ shimExprs ++ DayTimeIntervalShims.exprs ++ RoundingShims.exprs
   }
